Move onsuccess handler to the hook
